requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4

# Graph Modeling
networkx==3.2.1

# JSON Handling
orjson==3.9.10

# File Processing
python-magic==0.4.27

# Logging
colorlog==6.8.0

# Optional: For better PDF parsing
pdfplumber==0.10.3

# Optional: For enhanced text processing
spacy==3.7.2

